fixed resposive
